This is the import of the Soho Hotel theme as was delivered to us.
The theme might contain changes that diverge from upstream and we are
happy to revert those and make the master of this repository a more
clean and true version of the theme. In the same spirit we also
welcome contributions of the code fixes that has been done but not
publicly disclosed fixes of the theme - two security fixes among
others.